OBPIH-6500 Move invoice list page components into invoice/list directory
