+++ src/js/components/invoice/list/InvoiceListHeader.jsx
@@ -13,4 +13,3 @@ const InvoiceListHeader = () => (
 );
 
 export default withRouter(InvoiceListHeader);
-
